Move compute wer function to utils
