+++ src/speech_to_text_finetune/finetune_whisper.py
@@ -6,14 +6,12 @@ from transformers import (
     WhisperProcessor,
     WhisperForConditionalGeneration,
     Seq2SeqTrainingArguments,
-    EvalPrediction,
 )
 from transformers.models.whisper.english_normalizer import BasicTextNormalizer
 from transformers.models.whisper.tokenization_whisper import TO_LANGUAGE_CODE
 import torch
 from typing import Dict, Tuple
 import evaluate
-from evaluate import EvaluationModule
 from loguru import logger
 
 from speech_to_text_finetune.config import load_config
@@ -27,6 +25,7 @@ from speech_to_text_finetune.data_process import (
 from speech_to_text_finetune.utils import (
     get_hf_username,
     upload_custom_hf_model_card,
+    compute_wer_cer_metrics,
 )
 
 
@@ -180,69 +179,6 @@ def run_finetuning(
 
     logger.info(f"Find your final, best performing model at {local_output_dir}")
     return baseline_eval_results, eval_results
-
-
-def compute_wer_cer_metrics(
-    pred: EvalPrediction,
-    processor: WhisperProcessor,
-    wer: EvaluationModule,
-    cer: EvaluationModule,
-    normalizer: BasicTextNormalizer,
-) -> Dict:
-    """
-    Word Error Rate (wer) is a metric that measures the ratio of errors the ASR model makes given a transcript to the
-    total words spoken. Lower is better.
-    Character Error Rate (cer) is similar to wer, but operates on character instead of word. This metric is better
-    suited for languages with no concept of "word" like Chinese or Japanese. Lower is better.
-
-    More info: https://huggingface.co/learn/audio-course/en/chapter5/fine-tuning#evaluation-metrics
-
-    Note 1: WER/CER can be larger than 1.0, if the number of insertions I is larger than the number of correct words C.
-    Note 2: WER/CER doesn't tell the whole story and is not fully representative of the quality of the ASR model.
-
-    Args:
-        pred (EvalPrediction): Transformers object that holds predicted tokens and ground truth labels
-        processor (WhisperProcessor): Whisper processor used to decode tokens to strings
-        wer (EvaluationModule): module that calls the computing function for WER
-        cer (EvaluationModule): module that calls the computing function for CER
-        normalizer (BasicTextNormalizer): Normalizer from Whisper
-    Returns:
-        wer (Dict): computed WER metric
-    """
-
-    pred_ids = pred.predictions
-    label_ids = pred.label_ids
-
-    # replace -100 with the pad_token_id
-    label_ids[label_ids == -100] = processor.tokenizer.pad_token_id
-
-    # we do not want to group tokens when computing the metrics
-    pred_str = processor.batch_decode(pred_ids, skip_special_tokens=True)
-    label_str = processor.batch_decode(label_ids, skip_special_tokens=True)
-
-    # compute orthographic wer
-    wer_ortho = 100 * wer.compute(predictions=pred_str, references=label_str)
-    cer_ortho = 100 * cer.compute(predictions=pred_str, references=label_str)
-
-    # compute normalised WER
-    pred_str_norm = [normalizer(pred) for pred in pred_str]
-    label_str_norm = [normalizer(label) for label in label_str]
-    # filtering step to only evaluate the samples that correspond to non-zero references:
-    pred_str_norm = [
-        pred_str_norm[i]
-        for i in range(len(pred_str_norm))
-        if len(label_str_norm[i]) > 0
-    ]
-    label_str_norm = [
-        label_str_norm[i]
-        for i in range(len(label_str_norm))
-        if len(label_str_norm[i]) > 0
-    ]
-
-    wer = 100 * wer.compute(predictions=pred_str_norm, references=label_str_norm)
-    cer = 100 * cer.compute(predictions=pred_str_norm, references=label_str_norm)
-
-    return {"wer_ortho": wer_ortho, "wer": wer, "cer_ortho": cer_ortho, "cer": cer}
 
 
 if __name__ == "__main__":
